tt score cast fix
bench: 32500908

+++ src/TeaTable.cpp
@@ -44,7 +44,7 @@ void Teacup::update(Score s, NodeType nt, Depth d, Hash h, Move m, Index ply){
     if (s > MIN_VICTORY){ s += ply; }
     if (s < MIN_DEFEAT){ s -= ply; }
 
-    data |= (static_cast<uint64_t>(s) << 32); //do the mate score thing later
+    data |= (static_cast<uint64_t>(static_cast<uint16_t>(s)) << 32); //do the mate score thing later
     data |= (static_cast<uint64_t>(d) << 48);
     data |= (static_cast<uint64_t>(nt) << 54);
